requirements.txt
google-generativeai>=0.8.5
streamlit>=1.38.0
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
